ai_services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.1
pydantic==2.5.2
scikit-learn==1.3.2
python-dotenv==1.0.0
numpy==1.26.2
